+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 
 - Finish TTID span when transaction finishes (#3610)
 - Dont take screenshot and view hierarchy for app hanging (#3620)
-- Remove `free_storage` and `total_storage` from the app context (#3627), because Apple forbids sending
+- Remove `free_storage` and `storage_size` from the device context (#3627), because Apple forbids sending
 information retrieved via `NSFileSystemFreeSize` and `NSFileSystemSize` off a device; see
 [Apple docs](https://developer.apple.com/documentation/bundleresources/privacy_manifest_files/describing_use_of_required_reason_api?language=objc).